make D value readjustment work
Figured this out by examining the rsa.c for ffp-0.0.8, blast from the
past of 2003 and apparently the last time code for this technique was
written in a semi legible way.

The issue was we needed to recalculate dmp1 and dmq1, which I
mistakenly thought Go would do automatically once we adjusted D, but
the Precalculate() function short circuits if it detects any existing
values in the Precomputed struct.